refactor: split Vue2 app into components
